+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.1 under development
 
+- Bug #233: Refactor `DMLQueryBuilder`, related with yiisoft/db#746 (@Tigrov)
 - Enh #230: Improve column type #230 (@Tigrov)
 - Bug #240: Remove `RECURSIVE` expression from CTE queries (@Tigrov)